Move tests to Keboola\HttpExtractor\Tests namespace
